scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "pongpatapee", email = "pongpatapee1021@gmail.com"},
]
dependencies = [
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "pydantic>=2.9.2",
    "praw>=7.7.1",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
